Refactor Question
Question now contains the randomizeAnswerChoices() method which creates an array of strings out of the given incorrect answers and the correct answer and randomizes the placement of the correct answer.  It then returns the placement of the correct answer in the form of an index.

getCorrectAnswerIndex() returns that index for use in other classes.

getAnswers() returns the String array for use in other classes.