Add QuestionFormatter and QuestionFormatterTest and Refactor Main
QuestionFormatter has a method that formats the questions for printing in the console.

QuestionFormatterTest has a method called formatQuestionTest() that ensure the questions are formatted correctly.

Main was refactored to no longer do anything but run a console class that needs to be written.  This eliminates static calls as well as helps with MVS.